feat: rename `ext_manager.py` to `extmgr.py`; regard sub-directories as extensions
